requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Data Sources
yfinance==0.2.35
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Frontend
streamlit==1.30.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
